add Price Conversion functionality with Proper Validations. Price stored in the form of Indian Paise in Database

+++ server/models/productsTable.js
@@ -6,7 +6,7 @@ export async function createProductsTable() {
                 id UUID DEFAULT gen_random_uuid() PRIMARY KEY,
                 name VARCHAR(255) NOT NULL, 
                 description TEXT NOT NULL,
-                price DECIMAL(5,2) NOT NULL CHECK (price > 0),
+                price INT NOT NULL CHECK (price > 0),
                 category VARCHAR(100) NOT NULL,
                 ratings DECIMAL(3,2) DEFAULT 0 CHECK (ratings BETWEEN 0 AND 5),
                 images JSONB DEFAULT '[]'::JSONB,
